Movendo coisas csv-related para o arquivo csv.go
